+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## pre-release v0.7
 
 - Added support for AWS S3 buckets. Requires [ISHServer](https://github.com/Sarafian/ISHServer) version 1.1
+- [ISHServer](https://github.com/Sarafian/ISHServer) version 1.1 ISHCD provisioning is available. `Install-ISHDeployment.ps1` and `Uninstall-ISHDeployment.ps1` will calculate the cd's path.
 
 ## pre-release v0.6